requirements.txt
prometheus_client
python-dotenv
slick-bitcoinrpc
pycurl
